+++ apps/ledger-live-mobile/src/components/RootNavigator/SettingsNavigator.tsx
@@ -53,6 +53,7 @@ import { GenerateMockAccountSelectScreen } from "../../screens/Settings/Debug/Ge
 import HiddenNftCollections from "../../screens/Settings/Accounts/HiddenNftCollections";
 import { track } from "../../analytics";
 import { useCurrentRouteName } from "../../helpers/routeHooks";
+import PostOnboardingDebugScreen from "../../screens/PostOnboarding/PostOnboardingDebugScreen";
 
 // TODO: types for each screens and navigators need to be set
 export type SettingsNavigatorStackParamList = {
@@ -350,6 +351,10 @@ export default function SettingsNavigator() {
           headerShown: true,
           headerTitle: t("onboarding.stepLanguage.title"),
         }}
+      />
+      <Stack.Screen
+        name={ScreenName.PostOnboardingDebugScreen}
+        component={PostOnboardingDebugScreen}
       />
     </Stack.Navigator>
   );