+++ libs/ledgerjs/packages/hw-app-xrp/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-xrp
 
+## 6.27.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.27.4
+
+## 6.27.4-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.27.4-next.0
+
 ## 6.27.3
 
 ### Patch Changes